teams/trading_agents/requirements.txt
google-genai>=1.0.0
anthropic>=0.39.0
openai>=1.50.0
python-binance>=1.0.19
pandas>=2.0.0
ta>=0.11.0
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
pymongo>=4.6.0
pyyaml>=6.0
python-dotenv>=1.0
